requirements/development.txt
-r ./base.txt

pylint
django-debug-toolbar
pylint-django
isort
flake8